pico8: convert note::volume() to uint8_t so that maybe we can use bitfields later.

+++ src/pico8/memory.h
@@ -45,7 +45,7 @@ private:
 struct note
 {
     uint8_t key() const;
-    float volume() const;
+    uint8_t volume() const;
     uint8_t effect() const;
     uint8_t instrument() const;
